Trait interrupt_controller::AArch64LocalInterruptControllerApi

pub trait AArch64LocalInterruptControllerApi {
    // Required methods
    fn enable_fast_local_interrupt(&self, num: InterruptNumber, enabled: bool);
    fn send_fast_ipi(&self, num: InterruptNumber, dest: InterruptDestination);
    fn get_minimum_priority(&self) -> Priority;
    fn set_minimum_priority(&self, priority: Priority);
    fn acknowledge_interrupt(&self) -> Option<(InterruptNumber, Priority)>;
    fn init_secondary_cpu_interface(&self);
    unsafe fn acknowledge_fast_interrupt(
        &self
    ) -> Option<(InterruptNumber, Priority)>;
    unsafe fn end_of_fast_interrupt(&self, number: InterruptNumber);
}

AArch64-specific methods of a local interrupt controller

Required Methods

fn enable_fast_local_interrupt(&self, num: InterruptNumber, enabled: bool)

Same as LocalInterruptControllerApi::enable_local_interrupt but for fast interrupts (FIQs).

fn send_fast_ipi(&self, num: InterruptNumber, dest: InterruptDestination)

Same as LocalInterruptControllerApi::send_ipi but for fast interrupts (FIQs).

fn get_minimum_priority(&self) -> Priority

Reads the minimum priority for an interrupt to reach this CPU.

fn set_minimum_priority(&self, priority: Priority)

Changes the minimum priority for an interrupt to reach this CPU.

fn acknowledge_interrupt(&self) -> Option<(InterruptNumber, Priority)>

Aarch64-specific way to read the current pending interrupt number & priority.

fn init_secondary_cpu_interface(&self)

Aarch64-specific way to initialize the secondary CPU interfaces.

Must be called once from every secondary CPU.

unsafe fn acknowledge_fast_interrupt( &self ) -> Option<(InterruptNumber, Priority)>

Same as Self::acknowledge_interrupt but for fast interrupts (FIQs)

Safety

This is unsafe because it circumvents the internal Mutex. It must only be used by the interrupts crate when handling an FIQ.

unsafe fn end_of_fast_interrupt(&self, number: InterruptNumber)

Same as LocalInterruptControllerApi::end_of_interrupt but for fast interrupts (FIQs)

Safety

This is unsafe because it circumvents the internal Mutex. It must only be used by the interrupts crate when handling an FIQ.