Trait interrupt_controller::LocalInterruptControllerApi

pub trait LocalInterruptControllerApi {
    fn id(&self) -> LocalInterruptControllerId;
    fn get_local_interrupt_priority(&self, num: LocalInterruptNumber) -> Priority;
    fn set_local_interrupt_priority(
        &self,
        num: LocalInterruptNumber,
        priority: Priority
    );
    fn is_local_interrupt_enabled(&self, num: LocalInterruptNumber) -> bool;
    fn enable_local_interrupt(&self, num: LocalInterruptNumber, enabled: bool);
    fn send_ipi(&self, destination: InterruptDestination);
    fn get_minimum_priority(&self) -> Priority;
    fn set_minimum_priority(&self, priority: Priority);
    fn acknowledge_interrupt(&self) -> (LocalInterruptNumber, Priority);
    fn end_of_interrupt(&self, number: LocalInterruptNumber);
}

Required Methods

fn id(&self) -> LocalInterruptControllerId

fn get_local_interrupt_priority(&self, num: LocalInterruptNumber) -> Priority

fn set_local_interrupt_priority(
&self,
num: LocalInterruptNumber,
priority: Priority
)

fn is_local_interrupt_enabled(&self, num: LocalInterruptNumber) -> bool

fn enable_local_interrupt(&self, num: LocalInterruptNumber, enabled: bool)

fn send_ipi(&self, destination: InterruptDestination)

Sends an inter-processor interrupt to a specific CPU.

fn get_minimum_priority(&self) -> Priority

Reads the minimum priority for an interrupt to reach this CPU.

Note: aarch64-only, at the moment.

fn set_minimum_priority(&self, priority: Priority)

Changes the minimum priority for an interrupt to reach this CPU.

Note: aarch64-only, at the moment.

fn acknowledge_interrupt(&self) -> (LocalInterruptNumber, Priority)

Aarch64-specific way to read the current pending interrupt number & priority.

Always panics on x86_64.

fn end_of_interrupt(&self, number: LocalInterruptNumber)

Tell the interrupt controller that the current interrupt has been handled.